+++ rivr/tests.py
@@ -1,6 +1,5 @@
 from typing import Callable, Dict, Optional
-from rivr.request import Request
-from rivr.response import Response
+from rivr.http import Request, Response
 
 __all__ = ['TestClient']