+++ tests/test_router.py
@@ -2,8 +2,7 @@ import unittest
 
 from rivr.router import Router, Resolver404
 from rivr.views import View
-from rivr.request import Request
-from rivr.response import Response
+from rivr.http import Request, Response
 
 
 class RouterTest(unittest.TestCase):